Сформировал constants.js
